+++ src/libAtomVM/globalcontext.c
@@ -552,6 +552,23 @@ int globalcontext_get_registered_process(GlobalContext *glb, int atom_index)
     return 0;
 }
 
+term globalcontext_get_registered_name_process(GlobalContext *glb, int local_process_id)
+{
+    struct ListHead *registered_processes_list = synclist_wrlock(&glb->registered_processes);
+    struct ListHead *item;
+    struct ListHead *tmp;
+    MUTABLE_LIST_FOR_EACH (item, tmp, registered_processes_list) {
+        struct RegisteredProcess *registered_process = GET_LIST_ENTRY(item, struct RegisteredProcess, registered_processes_list_head);
+        if (registered_process->local_process_id == local_process_id) {
+            int result = registered_process->atom_index;
+            synclist_unlock(&glb->registered_processes);
+            return term_from_atom_index(result);
+        }
+    }
+    synclist_unlock(&glb->registered_processes);
+    return term_invalid_term();
+}
+
 bool globalcontext_is_atom_index_equal_to_atom_string(GlobalContext *glb, int atom_index_a, AtomString atom_string_b)
 {
     AtomString atom_string_a;